requirements.txt
Flask==2.3.3
requests==2.31.0
geopy==2.3.0
pytz==2023.3
python-docx==0.8.11
